evm: tidy codes unify signer functions in types
