+++ types/signer.go
@@ -1,0 +1,3 @@
+package types
+
+type Signer func(digest []byte) (signature []byte, err error)
